[#178] feat: 임명장 추가
