+++ src/components/results/ui/AlternativeBigCard.tsx
@@ -12,6 +12,21 @@ interface AlternativeCardUIProps {
   savingAmount: number;
 }
 
+const slideVariants = {
+  inactive: {
+    x: '-150%',
+  },
+  active: {
+    x: '-1000px',
+    transition: {
+      ease: [0.895, 0.03, 0.685, 0.22],
+      duration: 1,
+      repeat: Infinity,
+      repeatType: 'loop',
+    },
+  },
+};
+
 const AlternativeBigCard = ({ alternative, savingAmount }: AlternativeCardUIProps) => {
   const savingsPeriod = getSavingsPeriod(alternative.price, savingAmount);
